Add some presets, divide into several files for organization.
